+++ java-runtime-builder-app/src/main/java/com/google/cloud/runtimes/builder/buildsteps/MavenBuildStep.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.runtimes.builder.buildsteps.maven;
+package com.google.cloud.runtimes.builder.buildsteps;
 
 import com.google.cloud.runtimes.builder.buildsteps.base.AbstractSubprocessBuildStep;
 import com.google.cloud.runtimes.builder.buildsteps.base.BuildStepException;
