+++ java-runtime-builder-app/src/main/java/com/google/cloud/runtimes/builder/buildsteps/ScriptExecutionBuildStep.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.runtimes.builder.buildsteps.script;
+package com.google.cloud.runtimes.builder.buildsteps;
 
 import com.google.cloud.runtimes.builder.buildsteps.base.AbstractSubprocessBuildStep;
 import com.google.inject.Inject;
